refactor: relocate data modules
