+++ assets/data/attr_growth.js
