+++ assets/data/resources.js
