+++ assets/data/timing_engine.ts
@@ -5,7 +5,7 @@
    copy the needed types/guards or remove the imports and adapt signatures.
 */
 // import { Party, Member, TimedEffect, startChannel, stopChannel, recomputeResources } from "./party";
-// import { resolveSummon, SummonDef, SummonResolved } from "./assets/data/summons";
+// import { resolveSummon, SummonDef, SummonResolved } from "./summons";
 
 /* ========================= Core Types ========================= */
