+++ script.js
@@ -1,8 +1,8 @@
 import { SPELLBOOK } from "./assets/data/spells.js";
 import { WEAPON_SKILLS } from "./assets/data/weapon_skills.js";
-import { characterTemplate, gainProficiency } from "./core.js";
+import { characterTemplate, gainProficiency } from "./assets/data/core.js";
 import { getRaceStartingAttributes, RACE_DESCRIPTIONS } from "./assets/data/race_attrs.js";
-import { maxHP, maxMP, maxStamina } from "./resources.js";
+import { maxHP, maxMP, maxStamina } from "./assets/data/resources.js";
 import { DENOMINATIONS, CURRENCY_VALUES, convertCurrency, toIron, fromIron } from "./assets/data/currency.js";
 import { WEAPON_SLOTS, ARMOR_SLOTS, TRINKET_SLOTS } from "./assets/data/equipment.js";
